requirements.txt
Django
Pillow
Markdown
pip
